requirements.txt
# auto-trade-pro Requirements
poly-market-api>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
